+++ EasyToBuy.Repository/Concrete/CartRepository.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using EasyToBuy.Models.CommonModel;
+using EasyToBuy.Models.InputModels;
+using EasyToBuy.Models.Models;
+using EasyToBuy.Models.SPResults;
+using EasyToBuy.Repository.Abstract;
+using EasyToBuy.Services.Interactions;
+
+namespace EasyToBuy.Repository.Concrete
+{
+    public class CartRepository : ICartRepository
+    {
+        public async Task<ApiResponseModel> AddToCart(CartInputModel cartInputModel)
+        {
+            using (CartService cartService = new CartService())
+            {
+                return await cartService.AddToCart(cartInputModel);
+            }
+        }
+        public async Task<IEnumerable<SPGetCartDetailsByCustomerId_Result>> GetCartListByCustomerId(int customerId)
+        {
+            using (CartService cartService = new CartService())
+            {
+                return await cartService.GetCartListByCustomerId(customerId);
+            }
+        }
+        public async Task<ApiResponseModel> RemoveProductFromCart(int id)
+        {
+            using (CartService cartService = new CartService())
+            {
+                return await cartService.RemoveProductFromCart(id);
+            }
+        }
+    }
+}
