+++ EasyToBuy.Services/Interactions/ProductService.cs
@@ -61,130 +61,6 @@ namespace EasyToBuy.Services.Interactions
 
         #endregion
 
-        public async Task<IEnumerable<CategoryModel>> GetCategoryList()
-        {
-            var categoryList = new List<CategoryModel>();
-
-            try
-            {
-                var dbCategoryList = await _dbContext.tblCategory.ToListAsync();
-                foreach (var category in dbCategoryList)
-                {
-                    categoryList.Add(new CategoryModel
-                    {
-                        Id = category.Id,
-                        CategoryName = category.CategoryName,
-                        PackingMode = category.PackingMode,
-                        IsActive = category.IsActive,
-                    });
-                }
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-
-            return categoryList;
-        }
-        public async Task<IEnumerable<CategoryModel>> GetCategoryById(int Id)
-        {
-            var categoryById = new List<CategoryModel>();
-
-            try
-            {
-                var dbCategoryById = await _dbContext.tblCategory.Where(x => x.Id == Id).ToListAsync();
-                foreach (var category in dbCategoryById)
-                {
-                    categoryById.Add(new CategoryModel
-                    {
-                        Id = category.Id,
-                        CategoryName = category.CategoryName,
-                        PackingMode = category.PackingMode,
-                        IsActive = category.IsActive,
-                    });
-                }
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-
-            return categoryById;
-        }
-        public async Task<ApiResponseModel> CategoryAddEdit(CategoryInputModel categoryInputModel)
-        {
-            var apiResponseModel = new ApiResponseModel();
-            try
-            {
-                //var isCategoryExists = await _dbContext.tblCategory.Where(x => x.CategoryName == categoryInputModel.CategoryName && x.Id != categoryInputModel.Id).FirstOrDefaultAsync();
-
-                //if (isCategoryExists != null)
-                //{
-                //    apiResponseModel.Status = false;
-                //    apiResponseModel.Message = "This category already exists.";
-                //}
-                //else
-                //{
-                var dbCategory = await _dbContext.tblCategory.Where(x => x.Id == categoryInputModel.Id).FirstOrDefaultAsync();
-
-                if (dbCategory != null)
-                {
-                    dbCategory.CategoryName = categoryInputModel.CategoryName;
-                    dbCategory.PackingMode = categoryInputModel.PackingMode;
-                    dbCategory.UpdatedBy = categoryInputModel.UpdatedBy;
-                    dbCategory.UpdatedOn = DateTime.Now;
-                    dbCategory.IsActive = categoryInputModel.IsActive;
-                }
-                else
-                {
-                    var categoryObj = new Category();
-
-                    categoryObj.CategoryName = categoryInputModel.CategoryName;
-                    categoryObj.PackingMode = categoryInputModel.PackingMode;
-                    categoryObj.CreatedBy = categoryInputModel.CreatedBy;
-                    categoryObj.CreatedOn = DateTime.Now;
-                    categoryObj.IsActive = categoryInputModel.IsActive;
-
-                    await _dbContext.AddAsync(categoryObj);
-                }
-
-                await _dbContext.SaveChangesAsync();
-
-                apiResponseModel.Status = true;
-                apiResponseModel.Message = categoryInputModel.Id > 0 ? "Category updated successfully." : "Category added successfully.";
-                //}
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-            return apiResponseModel;
-        }
-        public async Task<ApiResponseModel> CategoryDelete(int Id)
-        {
-            var apiResponseModel = new ApiResponseModel();
-
-            try
-            {
-                var dbCategory = await _dbContext.tblCategory.FindAsync(Id);
-
-                if (dbCategory != null)
-                {
-                    dbCategory.IsActive = false;
-                    await _dbContext.SaveChangesAsync();
-
-                    apiResponseModel.Status = true;
-                    apiResponseModel.Message = "Category deleted successfully.";
-                }
-            }
-
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-
-            return apiResponseModel;
-        }
         public async Task<IEnumerable<ProductModel>> GetProductList()
         {
             var productList = new List<ProductModel>();
@@ -354,94 +230,6 @@ namespace EasyToBuy.Services.Interactions
 
             return apiResponseModel;
         }
-        public async Task<ApiResponseModel> AddToCart(CartInputModel cartInputModel)
-        {
-            var apiResponseModel = new ApiResponseModel();
-            try
-            {
-                var cartObj = new Cart();
-
-                cartObj.CustomerId = cartInputModel.CustomerId;
-                cartObj.ProductId = cartInputModel.ProductId;
-                cartObj.Quantity = cartInputModel.Quantity;
-                cartObj.AddedDate = DateTime.Now;
-
-                await _dbContext.AddAsync(cartObj);
-                await _dbContext.SaveChangesAsync();
-
-                apiResponseModel.Status = true;
-                apiResponseModel.Message = "Product added to cart successfully.";
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-            return apiResponseModel;
-        }
-        public async Task<IEnumerable<CartModel>> GetCartList()
-        {
-            var cartList = new List<CartModel>();
-
-            try
-            {
-                var dbCartList = await _dbContext.tblCart.ToListAsync();
-                foreach (var cart in dbCartList)
-                {
-                    cartList.Add(new CartModel
-                    {
-                        Id = cart.Id,
-                        CustomerId = cart.CustomerId,
-                        ProductId = cart.ProductId,
-                        Quantity = cart.Quantity,
-                        AddedDate = DateTime.Now,
-                    });
-                }
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-
-            return cartList;
-        }
-        public async Task<IEnumerable<SPGetCartDetailsByCustomerId_Result>> GetCartListByCustomerId(int customerId)
-        {
-            var cartListByCustomerId = new List<SPGetCartDetailsByCustomerId_Result>();
-
-            try
-            {
-                var sqlQuery = "exec spGetCartDetailsByCustomerId @CustomerId";
-
-                SqlParameter parameter = new SqlParameter("@CustomerId", customerId);
-
-                cartListByCustomerId = await _dbContext.cartDetailsByCustomerId_Results.FromSqlRaw(sqlQuery, parameter).ToListAsync();
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-
-            return cartListByCustomerId;
-        }
-        public async Task<ApiResponseModel> RemoveProductFromCart(int id)
-        {
-            var apiResponseModel = new ApiResponseModel();
-            try
-            {
-                var cartObj = await _dbContext.tblCart.Where(x => x.Id == id).FirstOrDefaultAsync();
-                if (cartObj != null)
-                {
-                    _dbContext.tblCart.Remove(cartObj);
-                    await _dbContext.SaveChangesAsync();
-                    apiResponseModel.Status = true;
-                    apiResponseModel.Message = "Product removed from cart successfully.";
-                }
-            }
-            catch (Exception ex)
-            {
-                var msg = ex.Message;
-            }
-            return apiResponseModel;
-        }
+ 
     }
 }
